backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1